fix borders in box.py and add border radius (rounding) feature
The code fixes the same issue that occured in circle.py 
e.g it sets a self._border_width var to the given value but doesnt do anything with it. The new code draws the border and the inner rectangle on top of it. 

Also a new feature added is the usage of a new var (self._border_radius), since pygame.draw.rect supports rounding a rectangle with a given radius it wasn't too hard to add. 

Signed-off-by: T <thollllllllllllllll@gmail.com>